oversampling first dataset

+++ multiqa/create_mixed_training.py
@@ -25,7 +25,7 @@ parser = argparse.ArgumentParser(description='Evaluate a model on TriviaQA data'
 parser.add_argument('datasets')
 parser.add_argument("--sample_first", type=float, default=1.0,
                         help="Percentage to sample first dataset")
-parser.add_argument("--sample_rest", type=int, default=1.0,
+parser.add_argument("--sample_rest", type=float, default=1.0,
                         help="Sample only this amount from training")
 args = parser.parse_args()
